+++ packages/net/src/http/index.ts
@@ -5,4 +5,5 @@
 export { IncomingMessage, IncomingMessageFromFetch } from './request.ts';
 export { ServerResponse } from './response.ts';
 export { HttpServer, createServer, request } from './server.ts';
+export { STATUS_CODES } from './status-codes.ts';
 export { default } from './server.ts';